feat: add photo and description
